requirements.txt
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0  #  delete
djangorestframework_simplejwt==4.3.0

mysql-connector-python==8.0.15
mysqlclient==1.3.12

requests==2.21.0

Faker==1.0.5
